Refactored into create_app() and blueprints
